added red separator lines on all screens where  the user is "editing" or adding shifts or employees. Trying to go for a theme to distinguish between pages that are some sort of configuration. Just to kind of draw the user's attention that they are changing something or capable of changing something on whatever page they are on. It also ties the whole app to a more uniform theme. Made all Navigation controller tintColor scheme the same to once again reinforce uniformity across the entire app.

+++ EditShiftDetailViewController.m
@@ -336,7 +336,7 @@ NSUInteger DeviceSystemMajorVersion()
     NSInteger ti = (NSInteger)interval;
     NSInteger minutes = (ti / 60) % 60;
     NSInteger hour = (ti / 3600);
-    return [NSString stringWithFormat:@"%i:%02i", hour, minutes];
+    return [NSString stringWithFormat:@"%li:%02li", (long)hour, (long)minutes];
 }
 
 /*! Adds or removes a UIDatePicker cell below the given indexPath.
